SD39.78 (#19)
Improvements to stretch.  If a call like pass in or
face in needs to know where the center of the whole set is,
that is is nominally improper with stretch.  But people call
things like stretch counter rotate, expecting it to be
changed to stretch split counter rotate.  The same will be
done with stretch pass in.  It will be permitted, but a
warning will be printed.
More improvements related to the 2024 Berkshires weekend.
Better handling of funny fractional circulates.
Handling of spread with NxN.
NxN pass and roll your neighbor.
Better handling of slim down.
Handling of parts of touch and go.
Better handling of melded.
Improvements in common spot.
Better handling of rectify 6x2 acey deucey.

Co-authored-by: Bill Ackerman <wba@alum.mit.edu>